Move IntoShares trait and friends to secret_sharing mod
Test client I am making to benchmark and execute different scenarios needs to be able to secret share user input before sending it to different helpers.

Helper server implementation does not need that, so trait is hidden behind a feature flag
