+++ src/test_fixture/circuit.rs
@@ -3,8 +3,8 @@ use crate::protocol::basics::SecureMul;
 use crate::protocol::context::Context;
 use crate::protocol::{QueryId, RecordId};
 use crate::rand::thread_rng;
-use crate::secret_sharing::Replicated;
-use crate::test_fixture::{narrow_contexts, share, Fp31, Reconstruct, TestWorld};
+use crate::secret_sharing::{share, Replicated};
+use crate::test_fixture::{narrow_contexts, Fp31, Reconstruct, TestWorld};
 use futures_util::future::join_all;
 
 /// Creates an arithmetic circuit with the given width and depth.
